# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JAMB Prep API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Subject {
  id: string;
  name: string;
  description: string;
  totalQuestions: number;
  icon: string;
}

export type QuestionCategory =
  (typeof QuestionCategory)[keyof typeof QuestionCategory];

export const QuestionCategory = {
  past: "past",
  repeated: "repeated",
  likely: "likely",
} as const;

export interface Question {
  id: string;
  subjectId: string;
  text: string;
  options: string[];
  correctAnswer: number;
  explanation: string;
  year?: number;
  category: QuestionCategory;
  topic: string;
}

export interface Note {
  id: string;
  subjectId: string;
  title: string;
  content: string;
  topic: string;
  order: number;
}

export type ProgressSubjectScores = { [key: string]: number };

export interface QuizResult {
  subjectId: string;
  score: number;
  total: number;
  timestamp: string;
  category: string;
}

export interface Progress {
  totalQuizzesTaken: number;
  totalCorrect: number;
  totalAttempted: number;
  subjectScores: ProgressSubjectScores;
  recentResults: QuizResult[];
}

export type GetQuestionsParams = {
  /**
   * Filter questions by category
   */
  category?: GetQuestionsCategory;
  /**
   * Limit number of questions
   */
  limit?: number;
};

export type GetQuestionsCategory =
  (typeof GetQuestionsCategory)[keyof typeof GetQuestionsCategory];

export const GetQuestionsCategory = {
  past: "past",
  repeated: "repeated",
  likely: "likely",
  all: "all",
} as const;