# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: JAMB Prep API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: subjects
    description: Subject operations
  - name: questions
    description: Question operations
  - name: progress
    description: Progress tracking
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /subjects:
    get:
      operationId: getSubjects
      tags: [subjects]
      summary: Get all subjects
      responses:
        "200":
          description: List of subjects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subject"

  /subjects/{subjectId}/questions:
    get:
      operationId: getQuestions
      tags: [questions]
      summary: Get questions for a subject
      parameters:
        - name: subjectId
          in: path
          required: true
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
            enum: [past, repeated, likely, all]
          description: Filter questions by category
        - name: limit
          in: query
          schema:
            type: integer
          description: Limit number of questions
      responses:
        "200":
          description: List of questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"

  /subjects/{subjectId}/notes:
    get:
      operationId: getNotes
      tags: [subjects]
      summary: Get study notes for a subject
      parameters:
        - name: subjectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Study notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"

  /progress:
    get:
      operationId: getProgress
      tags: [progress]
      summary: Get user progress
      responses:
        "200":
          description: User progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Progress"
    post:
      operationId: saveProgress
      tags: [progress]
      summary: Save quiz result
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizResult"
      responses:
        "200":
          description: Progress saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Progress"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Subject:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        totalQuestions:
          type: integer
        icon:
          type: string
      required:
        - id
        - name
        - description
        - totalQuestions
        - icon

    Question:
      type: object
      properties:
        id:
          type: string
        subjectId:
          type: string
        text:
          type: string
        options:
          type: array
          items:
            type: string
        correctAnswer:
          type: integer
        explanation:
          type: string
        year:
          type: integer
        category:
          type: string
          enum: [past, repeated, likely]
        topic:
          type: string
      required:
        - id
        - subjectId
        - text
        - options
        - correctAnswer
        - explanation
        - category
        - topic

    Note:
      type: object
      properties:
        id:
          type: string
        subjectId:
          type: string
        title:
          type: string
        content:
          type: string
        topic:
          type: string
        order:
          type: integer
      required:
        - id
        - subjectId
        - title
        - content
        - topic
        - order

    Progress:
      type: object
      properties:
        totalQuizzesTaken:
          type: integer
        totalCorrect:
          type: integer
        totalAttempted:
          type: integer
        subjectScores:
          type: object
          additionalProperties:
            type: number
        recentResults:
          type: array
          items:
            $ref: "#/components/schemas/QuizResult"
      required:
        - totalQuizzesTaken
        - totalCorrect
        - totalAttempted
        - subjectScores
        - recentResults

    QuizResult:
      type: object
      properties:
        subjectId:
          type: string
        score:
          type: integer
        total:
          type: integer
        timestamp:
          type: string
        category:
          type: string
      required:
        - subjectId
        - score
        - total
        - timestamp
        - category