# lib/api-zod/src/generated/types/note.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JAMB Prep API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Note {
  id: string;
  subjectId: string;
  title: string;
  content: string;
  topic: string;
  order: number;
}