# lib/api-zod/src/generated/types/progress.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JAMB Prep API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProgressSubjectScores } from "./progressSubjectScores";
import type { QuizResult } from "./quizResult";

export interface Progress {
  totalQuizzesTaken: number;
  totalCorrect: number;
  totalAttempted: number;
  subjectScores: ProgressSubjectScores;
  recentResults: QuizResult[];
}