# lib/api-zod/src/generated/types/question.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JAMB Prep API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QuestionCategory } from "./questionCategory";

export interface Question {
  id: string;
  subjectId: string;
  text: string;
  options: string[];
  correctAnswer: number;
  explanation: string;
  year?: number;
  category: QuestionCategory;
  topic: string;
}