# lib/api-zod/src/generated/types/quizResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JAMB Prep API specification
 * OpenAPI spec version: 0.1.0
 */

export interface QuizResult {
  subjectId: string;
  score: number;
  total: number;
  timestamp: string;
  category: string;
}