# lib/api-zod/src/generated/types/subject.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JAMB Prep API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Subject {
  id: string;
  name: string;
  description: string;
  totalQuestions: number;
  icon: string;
}